Move getCountriesPriceArray function to utils folder.
