+++ js/src/setup-mc/setup-stepper/setup-free-listings/shipping-rate/shipping-rate-setup/index.js
@@ -13,7 +13,7 @@ import AddRateButton from './add-rate-button';
 import CountriesPriceInputForm from './countries-price-input-form';
 import useStoreCurrency from '.~/hooks/useStoreCurrency';
 import useShippingRatesWithSavedSuggestions from './useShippingRatesWithSavedSuggestions';
-import getCountriesPriceArray from './getCountriesPriceArray';
+import getCountriesPriceArray from '.~/utils/getCountriesPriceArray';
 import AppSpinner from '.~/components/app-spinner';
 import useTargetAudienceFinalCountryCodes from '.~/hooks/useTargetAudienceFinalCountryCodes';
 import './index.scss';
